move callbacks 2 controller file
